Сделал стрелки у вектора { 11.02.24 }

+++ Point.h
@@ -2,7 +2,7 @@
 
 struct Point
 {
-	int x;
-	int y;
-	Point(int x = 0, int y = 0): x(x), y(y) {}
+	float x;
+	float y;
+	Point(float x = 0, float y = 0): x(x), y(y) {}
 };
